+++ packages/firebase_analytics/firebase_analytics_platform_interface/lib/src/method_channel/method_channel_firebase_analytics.dart
@@ -134,25 +134,6 @@ class MethodChannelFirebaseAnalytics extends FirebaseAnalyticsPlatform {
   }
 
   @override
-  Future<void> setCurrentScreen({
-    String? screenName,
-    String? screenClassOverride,
-    AnalyticsCallOptions? callOptions,
-  }) {
-    try {
-      return _api.logEvent(<String, Object?>{
-        'eventName': 'screen_view',
-        'parameters': <String, String?>{
-          'screen_name': screenName,
-          'screen_class': screenClassOverride,
-        },
-      });
-    } catch (e, s) {
-      convertPlatformException(e, s);
-    }
-  }
-
-  @override
   Future<void> setUserProperty({
     required String name,
     required String? value,