+++ packages/firebase_app_installations/firebase_app_installations/ios/firebase_app_installations/Package.swift
@@ -81,7 +81,7 @@ guard let shared_spm_version = Version("\(firebase_core_version_string)\(shared_
 let package = Package(
   name: "firebase_app_installations",
   platforms: [
-    .iOS("13.0"),
+    .iOS("15.0"),
   ],
   products: [
     .library(name: "firebase-app-installations", targets: ["firebase_app_installations"]),