+++ packages/firebase_auth/firebase_auth/lib/firebase_auth.dart
@@ -48,7 +48,6 @@ export 'package:firebase_auth_platform_interface/firebase_auth_platform_interfac
         YahooAuthProvider,
         YahooAuthCredential,
         MicrosoftAuthProvider,
-        MicrosoftAuthCredential,
         OAuthProvider,
         OAuthCredential,
         PhoneAuthProvider,