+++ packages/firebase_auth/firebase_auth_platform_interface/lib/src/method_channel/method_channel_user.dart
@@ -198,8 +198,6 @@ class MethodChannelUser extends UserPlatform {
                 androidPackageName: actionCodeSettings.androidPackageName,
                 androidInstallApp: actionCodeSettings.androidInstallApp,
                 androidMinimumVersion: actionCodeSettings.androidMinimumVersion,
-                // ignore: deprecated_member_use_from_same_package
-                dynamicLinkDomain: actionCodeSettings.dynamicLinkDomain,
                 linkDomain: actionCodeSettings.linkDomain,
               ),
       );
@@ -310,8 +308,6 @@ class MethodChannelUser extends UserPlatform {
                 androidPackageName: actionCodeSettings.androidPackageName,
                 androidInstallApp: actionCodeSettings.androidInstallApp,
                 androidMinimumVersion: actionCodeSettings.androidMinimumVersion,
-                // ignore: deprecated_member_use_from_same_package
-                dynamicLinkDomain: actionCodeSettings.dynamicLinkDomain,
                 linkDomain: actionCodeSettings.linkDomain,
               ),
       );