+++ packages/firebase_messaging/firebase_messaging_platform_interface/lib/src/platform_interface/platform_interface_messaging.dart
@@ -295,18 +295,6 @@ abstract class FirebaseMessagingPlatform extends PlatformInterface {
         'setForegroundNotificationPresentationOptions() is not implemented');
   }
 
-  /// Send a new [RemoteMessage] to the FCM server.
-  Future<void> sendMessage({
-    required String to,
-    Map<String, String>? data,
-    String? collapseKey,
-    String? messageId,
-    String? messageType,
-    int? ttl,
-  }) {
-    throw UnimplementedError('sendMessage() is not implemented');
-  }
-
   /// Subscribe to topic in background.
   ///
   /// [topic] must match the following regular expression: