+++ packages/firebase_ai/firebase_ai/CHANGELOG.md
@@ -1,6 +1,12 @@
-## 2.3.1
+## 3.1.0
 
  - **FEAT**(firebase_ai): Add support for `responseJsonSchema` in the [GenerationConfig].
+
+## 3.0.0
+
+> Note: This release has breaking changes.
+
+ - **BREAKING** **FEAT**: bump iOS SDK to version 12.0.0 ([#17549](https://github.com/firebase/flutterfire/issues/17549)). ([b2619e68](https://github.com/firebase/flutterfire/commit/b2619e685fec897513483df1d7be347b64f95606))
 
 ## 2.3.0