+++ packages/firebase_performance/firebase_performance_platform_interface/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.1.5+11
+
+ - Update a dependency to the latest release.
+
 ## 0.1.5+10
 
  - Update a dependency to the latest release.